Created TodoList.jsx and AddTodoForm.jsx, refactored App.jsx
